Edit:101

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -5,7 +5,7 @@
  */
 void print_number(int n)
 {
-	long m, num;
+	int m, num;
 	int c;
 	num = n;
